FOFBApp/Db and scripts/python: change offset from float to integer

+++ scripts/python/test_closed_loop.py
@@ -194,7 +194,7 @@ print('--------------------- Calculate the current offset ----------------------
 
 new_offset = np.zeros(channels)
 for i in range(0, channels):
-  new_offset[i] = np.mean(PV(pv_current_ArrayDataRAW[i]).get())
+  new_offset[i] = int(np.mean(PV(pv_current_ArrayDataRAW[i]).get()))
 
 print('New current offset values: \n', new_offset)
